+++ resources/views/components/layouts/main.blade.php
@@ -1,4 +1,8 @@
-<x-layouts.header />
+@props([
+    'metaRobots' => 'index, follow',
+])
+
+<x-layouts.header :metaRobots="$metaRobots" />
     <main id="main-content" class="sur-main flex min-h-0 flex-1 flex-col pt-20 outline-none lg:pt-[4.75rem]" tabindex="-1">
         {{ $slot }}
     </main>